pyproject.toml
[tool.poetry]
name = "spark"
version = "0.2.0"
description = "spark-on-k8s-repo"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
kubernetes = "23.3.0"
pyspark = "3.3.1"
delta-spark = "2.2.0"
numpy = "1.24.1"
pandas = "1.5.2"
pyyaml = "6.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"